@@ -6,7 +6,8 @@ import torch
 from omegaconf import OmegaConf as om
 
 from llmfoundry.models.layers import attention
-from llmfoundry.models.layers.attention import is_flash_v2_installed
+from llmfoundry.models.layers.attention import (check_alibi_support, gen_slopes,
+                                                is_flash_v2_installed)
 from llmfoundry.models.mpt.modeling_mpt import (apply_sequence_id,
                                                 gen_attention_mask_in_length,
                                                 gen_rotary_embedding)
@@ -20,7 +21,7 @@ def allclose_helper(t0: torch.Tensor,
 
 
 @pytest.mark.gpu
-@pytest.mark.parametrize('attn_impl_0,attn_impl_1', [
+@pytest.mark.parametrize('attn_impl_0, attn_impl_1', [
     ('flash', 'triton'),
     ('flash', 'torch'),
     ('triton', 'torch'),
@@ -74,9 +75,9 @@ def test_attn_impl(attn_impl_0: str,
     """
     alibi = pos_emb_config['alibi']
     rope = pos_emb_config['rope']
-    if alibi and (attn_impl_0 == 'flash' or attn_impl_1 == 'flash'):
-        pytest.skip('flash attn does not support alibi')
-
+    if alibi and not (check_alibi_support(attn_impl_0) and
+                      check_alibi_support(attn_impl_1)):
+        pytest.skip('flash attention below v2.4.2 does not support alibi.')
     if rope and (pos_emb_config['rope_impl']
                  == 'dail') and (not is_flash_v2_installed()):
         pytest.skip('dail implementation of rope requires flash attention 2.')
@@ -177,6 +178,12 @@ def test_attn_impl(attn_impl_0: str,
 
     with torch.autocast(x0.device.type):
         attn_bias_0 = gen_bias(attn_impl_0)
+        alibi_slopes_0 = None
+        if alibi and attn_impl_0 == 'flash':
+            alibi_slopes_0 = gen_slopes(n_heads=cfg.n_heads,
+                                        alibi_bias_max=8,
+                                        device=torch.device(device),
+                                        return_1d=True)
         rotary_emb_w_meta_info = None
         if rope:
             rotary_embedding = gen_rotary_embedding(
@@ -209,15 +216,23 @@ def test_attn_impl(attn_impl_0: str,
                          attention_mask=attention_mask,
                          rotary_emb_w_meta_info=rotary_emb_w_meta_info,
                          is_causal=True,
-                         attention_mask_in_length=attention_mask_in_length_0)
+                         attention_mask_in_length=attention_mask_in_length_0,
+                         alibi_slopes=alibi_slopes_0)
         attn_bias_1 = gen_bias(attn_impl_1)
+        alibi_slopes_1 = None
+        if alibi and attn_impl_1 == 'flash':
+            alibi_slopes_1 = gen_slopes(n_heads=cfg.n_heads,
+                                        alibi_bias_max=8,
+                                        device=torch.device(device),
+                                        return_1d=True)
         y1, _, _ = attn1(x1,
                          past_key_value=None,
                          attn_bias=attn_bias_1,
                          attention_mask=attention_mask,
                          rotary_emb_w_meta_info=rotary_emb_w_meta_info,
                          is_causal=True,
-                         attention_mask_in_length=attention_mask_in_length_1)
+                         attention_mask_in_length=attention_mask_in_length_1,
+                         alibi_slopes=alibi_slopes_1)
         y0 *= attention_mask.unsqueeze(-1)
         y1 *= attention_mask.unsqueeze(-1)
 
